Crate combinadics

This crate provides methods to de- and encode numbers in the combinatorial number system, sometime referred to as "combinadics".

Structs

DecodeIter
Iterates through all choices decoded from the given N and k in descending order.

Functions

decode
Returns an iterator over the choices decoded from the given N and k. The choices will be returned by the iterator in descending order.